move the set operation trait into a separate module
