+++ rust/healpix-geo-core/src/index/mod.rs
@@ -1,2 +1,3 @@
 mod indexers;
 mod region;
+mod set;
