+++ rust/healpix-geo-core/src/index/set.rs
@@ -1,0 +1,6 @@
+pub(crate) trait SetOperations {
+    fn union(&self, other: &Self) -> Self;
+    fn intersection(&self, other: &Self) -> Self;
+    fn difference(&self, other: &Self) -> Self;
+    fn symmetric_difference(&self, other: &Self) -> Self;
+}
